spikes/01_text_block_extraction/requirements.txt
PyMuPDF>=1.27.0
Pillow>=12.0.0
